fix: update AI service model to 'gpt-4.1', enhance network handling in transaction messages, and add utility functions for network and scan URL retrieval
