+++ src/shared/tools/tool-execution.ts
@@ -1,6 +1,7 @@
 import { TelegramBot } from '@/telegram-bot/telegram-bot';
 import { formatSmartNumber } from '@/telegram-bot/utils/format-text';
 import { EMessageType } from '../constants/enums';
+import { getScanUrl, getNetwork } from '../helper';
 
 /**
  * Enum representing the different states of a tool execution
@@ -182,15 +183,6 @@ export class ExampleToolExecutionCallback implements IToolExecutionCallback {
           data.toolName === ToolName.STAKE ||
           data.toolName === ToolName.TRANSFER)
       ) {
-        const getScanUrl = (network, txHash) => {
-          const scanUrls = {
-            bnb: `https://bscscan.com/tx/${txHash}`,
-            ethereum: `https://etherscan.io/tx/${txHash}`,
-            solana: `https://solscan.io/tx/${txHash}`,
-          };
-          return scanUrls[network] || `${txHash}`;
-        };
-
         let message;
 
         if (data.toolName === ToolName.SWAP) {
@@ -198,14 +190,15 @@ export class ExampleToolExecutionCallback implements IToolExecutionCallback {
           message = `🎉 <b>Congratulations, your transaction has been successful.</b>
 - <b>Swapped:</b> ${formatSmartNumber(data.data.fromAmount)} ${data.data.fromToken?.symbol || ''} 
 - <b>Received:</b> ${formatSmartNumber(data.data.toAmount)} ${data.data.toToken?.symbol || ''}
-- <b>Transaction Hash:</b> <a href="${scanUrl}">View on ${data.data.network.charAt(0).toUpperCase() + data.data.network.slice(1)} Explorer</a>
+- <b>Network:</b> ${getNetwork(data.data.network)}
+- <b>Transaction Hash:</b> <a href="${scanUrl}">View on ${getNetwork(data.data.network)} Explorer</a>
 `;
         } else if (data.toolName === ToolName.BRIDGE) {
           const scanUrl = getScanUrl(data.data.fromNetwork, data.data.transactionHash);
           message = `🎉 <b>Congratulations, your transaction has been successful.</b>
-- <b>Swapped:</b> ${formatSmartNumber(data.data.fromAmount)} ${data.data.fromToken?.symbol || ''} (${data.data.fromNetwork})
-- <b>Received:</b> ${formatSmartNumber(data.data.toAmount)} ${data.data.toToken?.symbol || ''} (${data.data.toNetwork})
-- <b>Transaction Hash:</b> <a href="${scanUrl}">View on ${data.data.fromNetwork.charAt(0).toUpperCase() + data.data.network.slice(1)} Explorer</a>
+- <b>Swapped:</b> ${formatSmartNumber(data.data.fromAmount)} ${data.data.fromToken?.symbol || ''} (${getNetwork(data.data.fromNetwork)})
+- <b>Received:</b> ${formatSmartNumber(data.data.toAmount)} ${data.data.toToken?.symbol || ''} (${getNetwork(data.data.toNetwork)})
+- <b>Transaction Hash:</b> <a href="${scanUrl}">View on ${getNetwork(data.data.fromNetwork)} Explorer</a>
 `;
         } else if (data.toolName === ToolName.STAKE) {
           const scanUrl = getScanUrl(data.data.network, data.data.transactionHash);
@@ -213,12 +206,14 @@ export class ExampleToolExecutionCallback implements IToolExecutionCallback {
           if (data.data.type === StakingOperationType.STAKE || data.data.type === StakingOperationType.SUPPLY) {
             message = `🎉 <b>Congratulations, your transaction has been successful.</b>
 - <b>Staked:</b> ${formatSmartNumber(data.data.amountA || 0)} ${data.data.tokenA?.symbol || ''}
-- <b>Transaction Hash:</b> <a href="${scanUrl}">View on ${data.data.network.charAt(0).toUpperCase() + data.data.network.slice(1)} Explorer</a>
+- <b>Network:</b> ${getNetwork(data.data.network)}
+- <b>Transaction Hash:</b> <a href="${scanUrl}">View on ${getNetwork(data.data.network)} Explorer</a>
           `;
           } else if (data.data.type === StakingOperationType.UNSTAKE || data.data.type === StakingOperationType.WITHDRAW) {
             message = `🎉 <b>Congratulations, your transaction has been successful.</b>
 - <b>Unstaked:</b> ${formatSmartNumber(data.data.amountA || 0)} ${data.data.tokenA?.symbol || ''}
-- <b>Transaction Hash:</b> <a href="${scanUrl}">View on ${data.data.network.charAt(0).toUpperCase() + data.data.network.slice(1)} Explorer</a>
+- <b>Network:</b> ${getNetwork(data.data.network)}
+- <b>Transaction Hash:</b> <a href="${scanUrl}">View on ${getNetwork(data.data.network)} Explorer</a>
           `;
           }
         } else if (data.toolName === ToolName.TRANSFER) {
@@ -226,7 +221,8 @@ export class ExampleToolExecutionCallback implements IToolExecutionCallback {
           message = `🎉 <b>Congratulations, your transaction has been successful.</b>
 - <b>To:</b> ${data.data.toAddress}
 - <b>Amount:</b> ${formatSmartNumber(data.data.amount || 0)} ${data.data.token?.symbol || ''}
-- <b>Transaction Hash:</b> <a href="${scanUrl}">View on ${data.data.network.charAt(0).toUpperCase() + data.data.network.slice(1)} Explorer</a>
+- <b>Network:</b> ${getNetwork(data.data.network)}
+- <b>Transaction Hash:</b> <a href="${scanUrl}">View on ${getNetwork(data.data.network)} Explorer</a>
         `;
         }
         this.messageData(EMessageType.TOOL_EXECUTION, message);
